pyproject.toml
[project]
name = "agentai"
version = "0.1.0"
description = "AgentAI MVP services (coordinator, executor, admin)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "greenlet>=3.1.1",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "redis>=5.0.8",
  "httpx>=0.27.0",
  "python-telegram-bot>=21.4",
  "PyYAML>=6.0.2",
  "python-frontmatter>=1.1.0",
  "prometheus-client>=0.20.0",
  "mem0ai>=0.1.0",
  "minio>=7.2.8",
  "kubernetes>=30.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "anyio>=4.4.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["services*", "libs*"]
